Minor changes (from float to double in some methods)

+++ src/main/java/eu/socialsensor/framework/client/dao/impl/KeywordDAOImpl.java
@@ -62,7 +62,7 @@ public class KeywordDAOImpl implements KeywordDAO {
 	}
     
     @Override
-    public void insertKeyword(String keyword, float score) {
+    public void insertKeyword(String keyword, double score) {
         Map<String, Object> map = new HashMap<String, Object>();
         String id = SocialNetworkSource.All+"::"+keyword;
         map.put("_id", id);
@@ -73,7 +73,7 @@ public class KeywordDAOImpl implements KeywordDAO {
     }
 
     @Override
-	public void insertKeyword(String keyword, float score, SocialNetworkSource sourceType) {
+	public void insertKeyword(String keyword, double score, SocialNetworkSource sourceType) {
     	Map<String, Object> map = new HashMap<String, Object>();
     	String id = sourceType+"::"+keyword;
         map.put("_id", id);
